alteration: modify login action

+++ src/server/controllers/users/Login.ts
@@ -2,14 +2,21 @@ import { Request, Response } from 'express';
 import { User } from '../../database/models';
 import { RefreshTokensProvider, UsersProvider } from '../../database/providers';
 import { StatusCodes } from 'http-status-codes';
-import { UnauthorizedError } from '../../utils/errors';
+import { ForbiddenError, UnauthorizedError } from '../../utils/errors';
 import { JWTService, PasswordService } from '../../shared/services';
+import { USER_STATUS } from '../../constants';
 
 
 
 export const login = async (req: Request<{}, {}, Pick<User, 'email' | 'password'>>, res: Response) => {
+	
 	const { email, password } = req.body;
 	const user = await UsersProvider.getByEmail(email);
+
+	if (user?.status === USER_STATUS.Inactive) { 
+		throw new ForbiddenError();
+	}
+
 	if (user && await PasswordService.equals(password, user.password)) {
 		const accessToken = JWTService.sign({ loggedUserId: user.id });
 		await RefreshTokensProvider.deleteByUserId(user.id);
@@ -26,5 +33,7 @@ export const login = async (req: Request<{}, {}, Pick<User, 'email' | 'password'
 				}
 			);
 	}
-	throw new UnauthorizedError('Incorrect e-mail or password');
+
+	throw new UnauthorizedError();
+
 };
